+++ src/main/java/tech/endorsed/signport/SignPort.java
@@ -6,6 +6,7 @@ import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import tech.endorsed.signport.command.AnchorCommand;
+import tech.endorsed.signport.config.SignPortConfig;
 import tech.endorsed.signport.events.SignEvents;
 
 public class SignPort implements ModInitializer {
@@ -14,6 +15,8 @@ public class SignPort implements ModInitializer {
 
 	@Override
 	public void onInitialize() {
+		SignPortConfig.load();
+
 		CommandRegistrationCallback.EVENT.register(
 				(dispatcher, registryAccess, world) -> {
 			AnchorCommand.register(dispatcher);