+++ src/main/java/tech/endorsed/signport/events/SignEvents.java
@@ -7,12 +7,10 @@ import net.minecraft.block.entity.SignBlockEntity;
 import net.minecraft.block.entity.SignText;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.text.Text;
-import net.minecraft.util.Pair;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 import tech.endorsed.signport.permission.SignPortPermissions;
-import tech.endorsed.signport.world.Anchor;
 import tech.endorsed.signport.world.PortSignEntity;
 
 public class SignEvents implements PlayerBlockBreakEvents.Before {
@@ -23,11 +21,7 @@ public class SignEvents implements PlayerBlockBreakEvents.Before {
 
         SignText front = ((SignBlockEntity) blockEntity).getFrontText();
         SignText back = ((SignBlockEntity) blockEntity).getBackText();
-        Pair<Boolean, Anchor> validPortSign = PortSignEntity.isValidPortSign(world, front);
-        if (!validPortSign.getLeft()) {
-            validPortSign = PortSignEntity.isValidPortSign(world, back);
-            if (!validPortSign.getLeft()) return true;
-        }
+        if (!PortSignEntity.isSignPortSign(front) && !PortSignEntity.isSignPortSign(back)) return true;
 
         player.sendMessage(Text.literal("You do not have permissions to remove port signs."), true);